Move AuditSzsLibraryEntry to dedicated utils file.
